talosip/src/requirements.txt
pycti==3.2.3
PyYAML==5.3.1
requests==2.23.0
wget